+++ src/main/java/com/CuteNekoDragon/Core/common/data/SVOTags.java
@@ -1,5 +1,6 @@
 package com.CuteNekoDragon.Core.common.data;
 
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
@@ -7,9 +8,12 @@ import net.minecraftforge.registries.ForgeRegistries;
 
 import com.CuteNekoDragon.Core.SVOCore;
 
+@SuppressWarnings("removal")
 public final class SVOTags {
 
     public static final class Items {
+
+        public static final TagKey<Item> HIDDEN_FROM_RECIPE_VIEWERS = TagKey.create(Registries.ITEM, new ResourceLocation("c", "hidden_from_recipe_viewers"));
 
         public static final TagKey<Item> Minerals = createItemTag("minerals");
         public static final TagKey<Item> Gemstones = createItemTag("gemstones");