+++ src/main/java/com/CuteNekoDragon/Core/common/datagen/recipes/SVORecipeProvider.java
@@ -1,11 +1,10 @@
-package com.CuteNekoDragon.Core.common.datagen;
+package com.CuteNekoDragon.Core.common.datagen.recipes;
 
+import com.CuteNekoDragon.Core.common.datagen.recipes.Shaped.ShapedRecipesProvider;
+import com.CuteNekoDragon.Core.common.datagen.recipes.Smithing.SmithingRecipesProvider;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.FinishedRecipe;
 import net.minecraft.data.recipes.RecipeProvider;
-
-import com.CuteNekoDragon.Core.common.datagen.recipes.ShapedRecipesProvider;
-import com.CuteNekoDragon.Core.common.datagen.recipes.SmithingRecipesProvider;
 
 import java.util.function.Consumer;